Proj Done !!!!

+++ src/com/sarvesh/hms/dto/Doctor.java
@@ -2,7 +2,7 @@ package com.sarvesh.hms.dto;
 
 public class Doctor {
 
-	private int id ;
+	private long id ;
 	private String userName ;
 	private String password ;
 	private String firstName ;
@@ -26,10 +26,10 @@ public class Doctor {
 	public void setSpecialization(String specialization) {
 		this.specialization = specialization;
 	}
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getUserName() {
@@ -93,6 +93,9 @@ public class Doctor {
 				+ ", bloodGroup=" + bloodGroup + ", gender=" + gender + ", degree=" + degree + ", specialization="
 				+ specialization + "]";
 	}
+	
+		
+
